% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se-accessors.R
\name{add_assay}
\alias{add_assay}
\title{Add assay to se}
\usage{
add_assay(se, new_assay, name, withDimnames = TRUE)
}
\arguments{
\item{se}{Summarized experiment}

\item{new_assay}{Assay to add as the main assay (first in the assays() list)}

\item{name}{Name of the added assay}

\item{withDimnames}{Logical; passed to \code{SummarizedExperiment::assays<-}.
If \code{TRUE} (default), the dimnames of the new assay are checked against
those of \code{se}.}
}
\value{
se with added assay. Old main assay is retained.
}
\description{
Convinience function to add a new assay directly as the main assay. The old assay is retained.
}
