% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se-construct.R
\name{add_randna}
\alias{add_randna}
\title{Explore type of missingness}
\usage{
add_randna(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
se with added rowData column
}
\description{
Add column to rowData that specifies for each row if values are missing at random or
not (MAR vs MNAR). The method is very rudimentary:
\enumerate{
\item Missing values are replaced by zeros, measured values by 1.
\item ANOVA between conditions is performed to determine significant differences
of the mean.
\item Rows with p-values < 0.05 are termed missing not at random.
}
}
