% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se-construct.R
\name{add_significance}
\alias{add_significance}
\title{Add Significance Information to a SummarizedExperiment Object}
\usage{
add_significance(se, p_thr = 0.05, diff_thr = 1)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{p_thr}{A numeric value for the p-value threshold (default: 0.05).}

\item{diff_thr}{A numeric value for the difference threshold (default: 1).}
}
\value{
A SummarizedExperiment object with added significance information.
}
\description{
This function takes a SummarizedExperiment object, calculates significance information,
and adds it to the rowData.
}
\examples{
\dontrun{
# Assuming 'se' is a SummarizedExperiment object
se_with_sign <- add_significance(se)
}
}
