% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{center_substring}
\alias{center_substring}
\title{Extract the centered substring of a given length from an input string}
\usage{
center_substring(input_string, n)
}
\arguments{
\item{input_string}{A character string from which the centered substring will be extracted.}

\item{n}{An integer specifying the length of the substring to be extracted.}
}
\value{
A character string representing the centered substring of the input string.
}
\description{
This function extracts a substring of length \code{n} from the center of the input string. If \code{n} is greater than the length of the input string,
the entire input string is returned.
}
\examples{
center_substring("hello", 3) # returns "ell"
center_substring("world", 2) # returns "or"
center_substring("example", 7) # returns "example"
}
