% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_values_colwise}
\alias{clean_values_colwise}
\title{Clean String Values Column-Wise}
\usage{
clean_values_colwise(x)
}
\arguments{
\item{x}{A character vector to be cleaned.}
}
\value{
A character vector with cleaned values.
}
\description{
This function cleans string values by:
\itemize{
\item Converting to lowercase.
\item Replacing all whitespace with underscores.
\item Replacing non-alphanumeric characters (except underscores) with underscores.
\item Collapsing multiple underscores into a single underscore.
\item Removing leading and trailing underscores.
\item Replacing the micro sign (Unicode U+00B5) with \code{u}, and notifying the user if replacements occur.
}
}
\examples{
library(dplyr)
library(stringr)

data <- data.frame(
  raw_column = c("10\u00b5g", "\u00b5mol/L", "Hello \u00b5World")
)

data_clean <- data \%>\%
  mutate(cleaned_column = clean_values_colwise(raw_column))

}
