% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-go.R
\name{enrich_go_se}
\alias{enrich_go_se}
\title{GO term enrichment}
\usage{
enrich_go_se(
  se,
  col_names = c(),
  simplify = TRUE,
  ont = "all",
  keyType = "SYMBOL",
  OrgDb = "org.Hs.eg.db",
  pvalueCutoff = 1
)
}
\arguments{
\item{se}{with t-test results (e.g. via DEP2 test_diff())}

\item{col_names}{name of columns on that enrichment is performed. If left empty, all columns ending with "_diff" are selected.}

\item{simplify}{Should similar terms be merged (sim > 0.7)}

\item{ont}{ontology}

\item{keyType}{Type of identifier}

\item{OrgDb}{Organism}

\item{pvalueCutoff}{significance threshold}
}
\value{
se with GO-term enrichment in metadata
}
\description{
Performs GO-term enrichment via clusterprofiler for all contrasts of a summarized experiment.
Results are added to the metadata.
}
