% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-proteomics.R
\name{fix_maxquant_pig_annotation}
\alias{fix_maxquant_pig_annotation}
\title{Handling of pig derived data}
\usage{
fix_maxquant_pig_annotation(
  proteingroups,
  peptides,
  fasta,
  mult_org = FALSE,
  obj = FALSE
)
}
\arguments{
\item{proteingroups}{String specifying path to ProteinGroups file}

\item{peptides}{String specifying path to ProteinGroups file}

\item{fasta}{String specifying path to ProteinGroups file}

\item{mult_org}{Logical specifying if fasta and experiment contain multiple organisms.}

\item{obj}{Logical specifying if results should be returned as one named list. If FALSE, results are saved as individual files.}
}
\value{
Either a named list containing all dataframes with correct protein names and gene names or locations where results have been stored.
}
\description{
Pig proteins are poorly annotated and have a different "style" when it comes to fasta files. Thus, the output of MaxQuant
is not directly usable. This function takes in a MaxQuant file from an experiment using pig derived proteins and a fasta
file and returns MaxQuant like files with correct protein names and gene names. In addition, columns are added that specify
the organism in case proteins from multiple organisms are expected. In this case, the fasta input must contain entries for
the other organisms as well in the "pig-style".
}
