% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-proteomics.R
\name{fragpipe_read_in}
\alias{fragpipe_read_in}
\title{fragpipe to se}
\usage{
fragpipe_read_in(
  file,
  gene_column = "gene",
  protein_column = "protein_id",
  sep = "_rep_",
  experimental_design = ""
)
}
\arguments{
\item{file}{path to proteinGroups or similar file}

\item{gene_column}{name of gene_name column after janitor}

\item{protein_column}{name of protein column after janitor}

\item{sep}{character describing the separator between sample name and replicate number (e.g. "\emph{rep}", "_r")}

\item{experimental_design}{dataframe with information regarding samples. If not specified,
sample names and groups are read automatically (start with "log2quantity_"), end with "_r" followed by
replicate number.}
}
\value{
summarized experiment
}
\description{
Reads in fragpipe or other output and creates a summarized experiment.
}
\details{
Mandatory columns: sample (sample name), condition (treatment), replicate
}
