% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-go.R
\name{genes_from_kegg}
\alias{genes_from_kegg}
\title{Retrieve Gene Data from a Specified KEGG Pathway}
\usage{
genes_from_kegg(keggpath_id = "hsa00190")
}
\arguments{
\item{keggpath_id}{A character string specifying the KEGG pathway ID.
Default is "hsa00190" for oxidative phosphorylation in Homo sapiens.}
}
\value{
A data frame containing Gene IDs, Descriptions, Gene symbols, KEGG ontology numbers,
Enzyme Commission numbers, the pathway ID queried, and the pathway name.
}
\description{
This function uses the R-package KEGGREST to retrieve data for a specified KEGG pathway and then returns a tidy data frame
that includes: "kegg_gene_id", "description", "gene_symbol", "kegg_ontology_number", "enzyme_commission_number", "kegg_path_number" and "kegg_path_name"
}
\examples{
\dontrun{
genes_from_kegg("hsa00190")
oxphos <- genes_from_kegg(keggpath_id = "hsa00190")  # default, or specify another pathway ID https://www.kegg.jp/entry/map00190
cellcycle <- genes_from_kegg(keggpath_id = "hsa04110") #https://www.genome.jp/pathway/map04110
mtor <- genes_from_kegg(keggpath_id = "hsa04150") #https://www.kegg.jp/pathway/map=map04150
ampk <- genes_from_kegg(keggpath_id = "hsa04152") #https://www.kegg.jp/pathway/map=map04152
print(cellcycle)


# Or mutliple keggpaths, assuming you have a vector of 4 keggpath_ids
keggpath_ids <- c("hsa00190", "hsa04110", "hsa04150", "hsa04152")

# Initialize an empty list to store the data frames
list_of_dataframes <- list()

# Loop over each keggpath_id
for (keggpath_id in keggpath_ids) {
  # Call the genes_from_kegg function for each keggpath_id
  df <- genes_from_kegg(keggpath_id)

  # Name the data frame according to the keggpath_id and add it to the list
  list_of_dataframes[[keggpath_id]] <- df
}

# Now, list_of_dataframes is a list of data frames, each named after a keggpath_id
# Use bind_rows to combine all data frames into a single data frame
combined_df <- bind_rows(list_of_dataframes, .id = "keggpath_id")
}
}
