% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_perseus}
\alias{impute_perseus}
\title{Impute as with perseus}
\usage{
impute_perseus(se, width = 0.3, downshift = 1.8, per_col = T)
}
\arguments{
\item{se}{summarized experiment}

\item{width}{width of gaussian distribution}

\item{downshift}{shift of gaussian distribution}

\item{per_col}{if distribution should be calculated per column (sample)
or accross all samples}
}
\value{
summarized experiment without missing values
}
\description{
Imputes missing values in a summarized experiment similar to Perseus. Missing values are
replaced by drawing from a left shifted gaussian distribution that is calculated based on either
individual samples or all measurements.
}
