% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isee.R
\name{isee_mini}
\alias{isee_mini}
\title{Create iSEE app with all custom panels}
\usage{
isee_mini(se)
}
\arguments{
\item{se}{summarized experiment object}
}
\value{
an iSEE app
}
\description{
Includes GeneAnnoTable, GOTable and RowDataTable.
}
