% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se-construct.R
\name{log2fc_to_fc}
\alias{log2fc_to_fc}
\title{Convert Log2 Differences to Fold Change in SummarizedExperiment Object}
\usage{
log2fc_to_fc(se)
}
\arguments{
\item{se}{A SummarizedExperiment object. This object should contain
rowData with columns that end with "_diff", representing log2 differences.}
}
\value{
A SummarizedExperiment object with the same structure, but with
additional columns where log2 values are converted to fold change.
}
\description{
This function takes a SummarizedExperiment object and adds new columns containing fold-changes.
The function looks for columns ending with "_diff" in the \code{rowData} slot and uses dplyr::mutate to 'unlog' them from log2 scale.
It preserves \code{NA} values in the input and returns a modified SummarizedExperiment object.
The function also checks to ensure that the rowData of the summarizedExperiment object remains
identical before and after the transformation, except for the newly created fold change columns.
Upon success, it prints a confirmation message and lists the names of the newly added columns.
}
\examples{
# Assuming `se` is a SummarizedExperiment object:
# se <- log2fc_to_fc(se)

}
