% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se-construct.R
\name{merge_se}
\alias{merge_se}
\title{merge_se}
\usage{
merge_se(se = list(), keep_all = FALSE)
}
\arguments{
\item{se}{A list of SummarizedExperiment objects}

\item{keep_all}{Logical indicating whether to keep all rows or only those present in all objects. Default is FALSE.}
}
\value{
A SummarizedExperiment object with the merged data.
}
\description{
merge_se
}
