% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-phospho.R
\name{phos_read_in_occ}
\alias{phos_read_in_occ}
\title{Read in phosphoproteomics occupancy data and process it}
\usage{
phos_read_in_occ(
  file,
  gene_column = "gene_names",
  protein_column = "proteins",
  sep = "_rep_",
  filt = c("reverse", "potential_contaminant"),
  keep_all_proteins = F,
  keep_all_genes = F,
  experimental_design = ""
)
}
\arguments{
\item{file}{A character string specifying the input file path.}

\item{gene_column}{A character string specifying the column name for gene names in the input file (default: "gene_names").}

\item{protein_column}{A character string specifying the column name for protein names in the input file (default: "proteins").}

\item{sep}{A character string specifying the separator for the replicate information in the column names (default: "\emph{rep}").}

\item{filt}{A character vector specifying the columns to filter for NA values (default: c("reverse", "potential_contaminant")).}

\item{keep_all_proteins}{A logical value indicating whether to keep all proteins when splitting protein groups (default: FALSE).}

\item{keep_all_genes}{A logical value indicating whether to keep all genes when splitting gene names (default: FALSE).}

\item{experimental_design}{A data frame specifying the experimental design. If not provided, it will be generated based on the input data.}
}
\value{
A SummarizedExperiment object containing the processed phosphoproteomics occupancy data.
}
\description{
This function reads in phosphoproteomics occupancy data from the "Phospho (STY)Sites.txt" MaxQuant output,
processes the data, filters false and low-quality hits, and returns a SummarizedExperiment object.
}
\examples{
\dontrun{
# Example input file path (replace with your own file path)
input_file <- "path/to/your/input_file.tsv"

# Read and process the data
data_se <- phos_read_in_occ(input_file)
}
}
