% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-go.R
\name{phospho_ora}
\alias{phospho_ora}
\title{Perform Over-representation Analysis (ORA) on Phosphoproteomics Data}
\usage{
phospho_ora(
  se,
  contr = "all",
  OrgDb = "org.Hs.eg.db",
  pvalueCutoff = 0.4,
  qvalueCutoff = 0.8,
  ont = c("BP", "MF", "CC")
)
}
\arguments{
\item{se}{A SummarizedExperiment object containing phosphoproteomics data.}

\item{contr}{A character vector specifying the contrasts to be analyzed. Default is "all".}

\item{OrgDb}{A character string specifying the organism database to be used. Default is "org.Hs.eg.db".}

\item{pvalueCutoff}{Numeric value for the p-value cutoff. Default is 0.4.}

\item{qvalueCutoff}{Numeric value for the q-value cutoff. Default is 0.8.}

\item{ont}{A character vector specifying the Gene Ontology (GO) categories to be analyzed. Default is c("BP", "MF", "CC").}
}
\value{
A list containing the ora object, a combined dataframe and a ggplot DEP2icting terms with q < 0.2
}
\description{
This function takes a SummarizedExperiment object and conducts Over-representation Analysis (ORA) on proteins with at least one significant phosphosite, using the clusterProfiler package.
}
