% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline_result}
\alias{pipeline_result}
\title{The object a pipeline produced}
\usage{
pipeline_result(pipe)
}
\arguments{
\item{pipe}{A \code{proteoSE_pipeline} that has been run.}
}
\value{
The result of the final step.
}
\description{
The object a pipeline produced
}
\examples{
pipe <- run_pipeline(add_step(proteoSE_pipeline(1:10), "rev"))
pipeline_result(pipe)
}
