% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap_clustered}
\alias{plot_heatmap_clustered}
\title{Create a clustered heatmap from a given SummarizedExperiment object}
\usage{
plot_heatmap_clustered(
  se,
  indicate = "condition",
  type = "centered",
  k = 3,
  ...
)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{indicate}{A character string indicating the variable to be used for annotation.}

\item{type}{A character string indicating the type of heatmap.}

\item{k}{An integer specifying the number of clusters to be used in k-means clustering.}

\item{...}{Additional arguments to be passed to the underlying ComplexHeatmap::Heatmap function.}
}
\value{
A list containing the heatmap plot, a data frame with the heatmap data, and a data frame with the protein clusters.
}
\description{
Create a clustered heatmap from a given SummarizedExperiment object
}
\examples{
\dontrun{
# Assuming se_diff is a SummarizedExperiment object with appropriate data
result <- plot_heatmap_clustered(se_diff, indicate = "condition", type = "centered", k = 5)
result$plot
result$df
result$clusters
}
}
