% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Create a 2D plot_scatter}
\usage{
plot_scatter(
  df,
  col_x,
  col_y,
  col_label = "gene_names",
  show_labels = TRUE,
  title = "",
  standard_dev = 2,
  window = 1
)
}
\arguments{
\item{df}{data for plotting}

\item{col_x}{string column name of x-axis data}

\item{col_y}{string column name of y-axis data}

\item{col_label}{string column name of labels}

\item{show_labels}{Boolean show labels on plot}

\item{title}{string plot title}

\item{standard_dev}{int difference to sd that specifies highlighted entries}

\item{window}{rolling window for which sd is calculated}
}
\value{
list containing the scatterplot and dfs for highlighted entries
}
\description{
Plots two columns against each other and marks entries that differ more than standard_dev SD. This calculation is performed as a moving window.
}
