% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosr.R
\name{prep_phosR}
\alias{prep_phosR}
\title{Prepares Phosphorylation Data for Analysis}
\usage{
prep_phosR(se, species = "human", numMotif = 5, numSub = 1, top = 30)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{species}{A character string specifying the species (default is "human").
Must be one of "human", "mouse", or "rat".}

\item{numMotif}{An integer specifying the number of motifs (default is 5).}

\item{numSub}{An integer specifying the number of substrates (default is 1).}

\item{top}{An integer specifying the top number of substrates to select (default is 30).}
}
\value{
A list containing the following elements:
- kinaseSubstrateScore: A list of kinase substrate scores.
- kinaseSubstratePred: A matrix of kinase substrate predictions.
- mat: A matrix of collapsed phosphorylation data.
- kinase_all_color: A color palette for plotting.
- kinase_signalome_color" = A color palette for plotting.
- seq: A vector of sequences.
}
\description{
This function takes a SummarizedExperiment object and prepares the data for
phosphorylation analysis using the PhosR package. It also generates a color palette
for plotting.
}
