% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosr.R
\name{prep_phosR_from_ppe}
\alias{prep_phosR_from_ppe}
\title{Prepares phosphoproteomics data for PhosR analysis}
\usage{
prep_phosR_from_ppe(
  ppe,
  species = "human",
  numMotif = 5,
  numSub = 1,
  top = 30,
  assay = "z_scored"
)
}
\arguments{
\item{ppe}{Phosphoproteomics data object}

\item{species}{Species of the data, one of "human", "mouse", or "rat" (default: "human")}

\item{numMotif}{Number of motifs to use (default: 5)}

\item{numSub}{Number of substrates to use (default: 1)}

\item{top}{Number of top kinases to consider (default: 30)}

\item{assay}{Type of assay data to use (default: "z_scored")}
}
\value{
A list containing kinaseSubstrateScore, kinaseSubstratePred, mat, kinase_all_color, kinase_signalome_color, and seq
}
\description{
This function processes phosphoproteomics data, calculates kinase-substrate scores, and generates predictions for the given data.
It also generates color palettes for visualization.
}
