% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-gsea.R
\name{prep_ssgsea2}
\alias{prep_ssgsea2}
\title{Prepare Data for Single Sample Gene Set Enrichment Analysis (ssGSEA)}
\usage{
prep_ssgsea2(se, file = "")
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{file}{A character string indicating the file name to save the GCT object. Defaults to an empty string, which means the GCT object will not be saved to a file.}
}
\value{
A list with the following elements:
\describe{
\item{"data"}{A GCT object with the unique rows from the input SummarizedExperiment object.}
\item{"path"}{A character string indicating the path of the saved GCT file, if applicable.}
\item{"duplicated"}{A SummarizedExperiment object containing the duplicated rows removed from the input object.}
}
}
\description{
This function prepares the data from a SummarizedExperiment object for ssGSEA analysis. It removes duplicated rows based on the rowData sequence_window, creates a GCT object, and optionally writes the GCT object to a file.
}
\examples{
\dontrun{
# Create a SummarizedExperiment object (se) here
# ...
result <- prep_ssgsea2(se)
result <- prep_ssgsea2(se, file = "output")
}
}
