% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{proteoSE-deprecated-ptm}
\alias{proteoSE-deprecated-ptm}
\alias{maxq_to_ppe}
\alias{ptm_to_ppe}
\title{Deprecated PTM-to-PhosphoExperiment converters}
\usage{
maxq_to_ppe(...)

ptm_to_ppe(...)
}
\arguments{
\item{...}{Passed on to \code{\link{read_ptm_to_ppe}}.}
}
\value{
A \code{\link[PhosR]{PhosphoExperiment}} object (see
\code{\link{read_ptm_to_ppe}}).
}
\description{
These functions are deprecated. Use \code{\link{read_ptm_to_ppe}} instead.
}
\keyword{internal}
