% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isee.R
\name{se_to_isee}
\alias{se_to_isee}
\title{Make se compatible to iSEE}
\usage{
se_to_isee(se, PValuePatterns = "p.val", LogFCPatterns = "_diff")
}
\arguments{
\item{se}{summarized experiment}

\item{PValuePatterns}{Character vector of column-name patterns identifying
p-value columns to register for volcano plots (default \code{"p.val"}).}

\item{LogFCPatterns}{Character vector of column-name patterns identifying
log fold-change columns to register for volcano plots (default \code{"_diff"}).}
}
\value{
se summarized experiment with registered features
}
\description{
Transform summarized experiment. The result is a summarized experiment
of the type "SingleCellExperiment" to allow dimensional reduction plots.
GO-enrichment is registered from the metadata as a feature set collection and
p-values and differences are registered to be plotted via volcano plots.
#'
}
