% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sev-package.R
\docType{package}
\name{sev-package}
\alias{sev}
\alias{sev-package}
\title{sev: The SummarizedExperiment Viewer for Proteomics Analysis}
\description{
A utility toolkit for (phospho)proteomics data analysis organised around the
Bioconductor \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} class. It provides
convenience functions for data import, missing-value imputation,
differential-abundance testing, GO-term and gene-set enrichment, PhosR-based
signalome analysis, STRING network retrieval and publication-ready plotting,
built on top of DEP2, iSEE/iSEEu, PhosR, limma and clusterProfiler.
}
\details{
The package source is grouped by theme (see the file names in \verb{R/}). For an
overview of the refactoring history and outstanding cleanup tasks, see
\code{AUDIT.md} and \verb{CHANGES_phase0-1.md} in the package root.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adhutz/sev}
  \item Report bugs at \url{https://github.com/adhutz/sev/issues}
}

}
\author{
\strong{Maintainer}: Arvid Hutzfeldt \email{arvid.hutzfeldt@gmail.com}

}
\keyword{internal}
