% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-test.R
\name{subset_results}
\alias{subset_results}
\title{Subset a Dataframe Based on Conditions, Contrasts, and Additional Columns}
\usage{
subset_results(
  df,
  id_col = "name",
  condition_list,
  contrast_list,
  get_summaries = TRUE,
  extra_cols = NULL
)
}
\arguments{
\item{df}{A dataframe from which to subset.}

\item{id_col}{A string specifying the name of the ID column. Default is "name".}

\item{condition_list}{A character vector of conditions. Columns matching these conditions followed by numbers
will be selected.}

\item{contrast_list}{A character vector of contrasts. Columns starting with these contrasts will be selected.}

\item{get_summaries}{A logical value indicating whether to include summary intensity columns for the specified conditions.
The intensity summary columns include mean, standard deviation, min, max, quartiles, IQR, missing counts, and measured counts.
Default is TRUE.}

\item{extra_cols}{A character vector of additional column names to include in the subset. The columns must exist in the dataframe.
Default is NULL, meaning no extra columns will be selected.}
}
\value{
A dataframe containing the selected columns in the following order: ID column, extra columns (if specified),
condition columns (including intensity summary columns if \code{get_summaries} is TRUE in the specified order), and contrast columns.
}
\description{
This function subsets a dataframe by selecting specified condition columns, contrast columns,
and optionally extra columns. It also supports selecting summary intensity columns for conditions,
and ensures that the intensity summary columns are selected in a specific order.
}
\examples{
# Example dataframe
df <- data.frame(
  name = c("A", "B", "C"),
  cond1_1 = c(1, 2, 3),
  cond1_mean_intensity = c(10, 20, 30),
  contrastA_1 = c(100, 200, 300),
  col2 = c("X", "Y", "Z")
)

# Subset the dataframe with conditions and contrasts
subset_results(df, id_col = "name", condition_list = c("cond1"), contrast_list = c("contrastA"), extra_cols = c("col2"))

}
