% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-test.R
\name{test_diff_long}
\alias{test_diff_long}
\title{Long Test}
\usage{
test_diff_long(se)
}
\arguments{
\item{se}{An se object.}
}
\value{
A data frame.
}
\description{
This function returns test results for an se object in long format.
}
\examples{
# Example usage:
# result <- test_diff_long(se)
}
