% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_sev}
\alias{theme_sev}
\title{Custom theme for ggplot2}
\usage{
theme_sev()
}
\value{
A ggplot2 theme object.
}
\description{
This function applies a custom theme to a plot by combining the DEP2::theme_DEP1()
function with additional theme elements.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
p + theme_sev()
}
