% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{write_phos}
\alias{write_phos}
\title{Write phosphoproteomics data to a file}
\usage{
write_phos(se, file = "")
}
\arguments{
\item{se}{A SummarizedExperiment object containing phosphoproteomics data.}

\item{file}{The filename where the resulting data should be saved (default: current directory).}
}
\value{
relevant data
}
\description{
This function takes a SummarizedExperiment object, extracts the relevant
data, and writes it to a file.
}
