% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{condition_impute}
\alias{condition_impute}
\title{Condition-wise Imputation for Proteomics Data}
\usage{
condition_impute(
  se,
  fun = c("QRILC", "bpca", "knn", "MLE", "MinDet", "MinProb", "man", "min", "zero",
    "mixed", "nbavg", "RF", "GSimp"),
  percent = 0.3,
  percent_type = c("measured", "NA"),
  logical_operator = ">",
  ...
)
}
\arguments{
\item{se}{SummarizedExperiment
Proteomics data in a \code{\link[SummarizedExperiment]{SummarizedExperiment}} object.}

\item{fun}{Character
Imputation method to use. Available options include:
\code{"QRILC"}, \code{"bpca"}, \code{"knn"}, \code{"MLE"}, \code{"MinDet"}, \code{"MinProb"},
\code{"man"}, \code{"min"}, \code{"zero"}, \code{"mixed"}, \code{"nbavg"}, \code{"RF"}, \code{"GSimp"}.
See \code{\link[MSnbase]{impute,MSnSet-method}}, \code{\link[DEP2]{manual_impute}}, and
\code{\link[missForest]{missForest}} for details.}

\item{percent}{Numeric
A value between 0 and 1 specifying the percentage threshold for filtering rows based
on measured or missing values.}

\item{percent_type}{Character
Specifies whether to filter rows based on the percentage of measured (\code{"measured"}) or
missing (\code{"NA"}) values.}

\item{logical_operator}{Character
Logical operator to apply for filtering rows. Options are: \code{">"}, \code{"<"},
\code{">="}, \code{"<="}, \code{"=="}, \code{"!="}.}

\item{...}{Additional arguments for specific imputation methods. For example:
\itemize{
\item \code{man}: \code{\link[DEP2]{manual_impute}} supports \code{shift} and \code{scale}.
\item \code{RF}: \code{\link[missForest]{missForest}} supports \code{ntree} and \code{parallelize}.
\item Other methods: See \code{\link[MSnbase]{impute,MSnSet-method}}.
}}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object with imputed values.
}
\description{
\code{condition_impute} imputes missing values in a proteomics dataset on a condition-wise basis.
It supports various imputation methods and allows filtering rows based on the percentage
of measured or missing values, with customizable logical operators.
}
\examples{
\dontrun{
# Example usage with SummarizedExperiment data
# Impute NA in rows where more than 30\% of values are measured
imputed_se <- condition_impute(se, 
                               fun = "man", 
                               percent = 0.3, 
                               percent_type = "measured", 
                               logical_operator = ">",
                               shift = 1.8, scale = 0.3)

# Impute NA in rows where more than 70\% of values are missing
imputed_se <- condition_impute(se, 
                               fun = "knn", 
                               percent = 0.7, 
                               percent_type = "NA", 
                               logical_operator = ">",
                               k = 5, rowmax = 0.8)

# Impute NA in rows where missing values are <=50\%
imputed_se <- condition_impute(se, 
                               fun = "QRILC", 
                               percent = 0.5, 
                               percent_type = "NA", 
                               logical_operator = "<=")

}
}
