% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-antigen.R
\name{plot_antigen}
\alias{plot_antigen}
\title{Plot Antigen Intensities and Differences}
\usage{
plot_antigen(
  se,
  contrast,
  additional_sets = "none",
  scale = FALSE,
  min_diff = 1,
  min_intensity = 10,
  max.overlaps = 40,
  targets = data.frame(name = character(0), target = character(0), min.segment.length =
    0)
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object containing the experimental data.}

\item{contrast}{A string specifying the comparison between conditions in the format "test_vs_control".}

\item{additional_sets}{A character vector specifying additional condition sets to include in the analysis. Defaults to "none", meaning no additional sets are included. Set to "all" to include all conditions.}

\item{scale}{A logical value indicating whether to scale the data by z-transformation. Defaults to \code{FALSE}.}

\item{min_diff}{A numeric value specifying the minimum difference required for data points to be considered. Defaults to \code{1}.}

\item{min_intensity}{A numeric value specifying the minimum intensity for data points to be considered. Defaults to \code{10}. If scaled is set to TRUE, this needs to be changed to a lower value.}

\item{max.overlaps}{A numeric value indicating the maximum number of label overlaps allowed in the plot. Defaults to \code{40}.}

\item{targets}{A data frame containing \code{name} and \code{target} columns, specifying target proteins or features of interest. Defaults to an empty data frame.}

\item{min.segment.length}{Minimum segment length for labels}
}
\value{
A combined plot displaying individual and overlapping data points for the specified contrast.
}
\description{
This function creates patient-wise plots of protein intensity differences for a given contrast between conditions in a \code{SummarizedExperiment} object. Optionally, intensities can be z-scaled. For z-score calculation, additional conditions can be included. The function also generates an overlap plot that depicts in how many patients the given proteins were above the thresholds.
}
\details{
The function processes the \code{SummarizedExperiment} data to calculate the mean of control samples and the differences between test samples and control means. It generates individual plots for test samples and an overlap plot for the specified conditions.

If the mean for the control condition is not already present in the data, it is calculated using \code{sev2::add_stats()}.
}
\examples{
\dontrun{
# Assuming `se` is a valid SummarizedExperiment object and `contrast` is specified:
plot_antigen(se, contrast = "treatment_vs_control")

}
}
