% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-antigen.R
\name{plot_antigen_missing}
\alias{plot_antigen_missing}
\title{Plot Antigen Missing Data Visualization}
\usage{
plot_antigen_missing(
  se,
  test_condition = "lcm_igan",
  ctrl_condition = "lcm_ctrl",
  quantile = 0.1,
  perc_low_ctrl = 80,
  targets = data.frame(name = character(0), target = character(0))
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object containing the data to be analyzed.}

\item{test_condition}{A string specifying the test condition to analyze. Defaults to \code{"lcm_igan"}.}

\item{ctrl_condition}{A string specifying the control condition to compare against. Defaults to \code{"lcm_ctrl"}.}

\item{quantile}{A numeric value indicating the quantile cutoff for low intensity in the control condition. Defaults to \code{0.1}.}

\item{perc_low_ctrl}{A numeric threshold for filtering antigens based on the percentage of low-intensity measurements in the control condition. Defaults to \code{80}.}

\item{targets}{A data frame with \code{name} and \code{target} columns to highlight specific targets. Defaults to an empty data frame.}
}
\value{
A \code{ggplot2} object showing a plot with points representing antigens, color-coded based on their missing data percentages, and optionally labeled.
}
\description{
This function creates a detailed plot to visualize missing data patterns, highlighting the percentage of missing data in test and control conditions. It also allows for the labeling of specific targets.
}
\details{
The function calculates missing and measured data percentages, applies quantile-based cutoffs for low intensity, and generates a scatter plot with facets showing different categories of missing data in the test and control conditions. The visualization helps identify patterns of data availability.
}
\examples{
\dontrun{
# Assuming `se` is a `SummarizedExperiment` object with relevant data:
plot_antigen_missing(se, test_condition = "test_cond", ctrl_condition = "control_cond", quantile = 0.05, perc_low_ctrl = 70)

}
}
